requirements.txt
attrs==21.4.0
autopep8==1.6.0
Bootstrap-Flask==2.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coverage==6.3.3
dependency-injector==4.39.1
Deprecated==1.2.13
Flask==2.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
PyGithub==1.55
PyJWT==2.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-flask==1.2.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
